requirements.txt
opencv-python
requests
python-dotenv
flask
flask_migrate
flask_sqlalchemy
#python -m venv venv
# venv\Scripts\activate
# pip install -r requirements.txt
